refactored code a little more
